main/requirements.txt
pillow
pytesseract
wand
beautifulsoup4
imutils
opencv-python